marvexbe/requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
csvkit==1.0.3
dbfread==2.0.7
diff-match-patch==20121119
Django==1.11.11
django-filter==1.1.0
django-filters==0.2.1
django-import-export==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
et-xmlfile==1.0.1
future==0.16.0
idna==2.6
isodate==0.6.0
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
leather==0.3.3
MarkupSafe==1.0
MySQL-python==1.2.5
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.1
parsedatetime==2.4
pkg-resources==0.0.0
python-slugify==1.2.4
pytimeparse==1.1.7
pytz==2018.3
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.5
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
